# api-doc.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: VinDr Lab API Document
  contact:
    name: VinDr Lab Development Team
  license:
    name: MIT
paths:
  /labels:
    get:
      description: get Labels by array
      operationId: fetchLabels
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: label_group_id
          in: query
          required: false
          style: form
          schema:
            type: array
            items:
              type: string
              format: uuid
        - name: project_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          impression:
                            type: array
                            items:
                              $ref: "#/components/schemas/Label"
                          finding:
                            type: array
                            items:
                              $ref: "#/components/schemas/Label"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      description: |
        create a new Label

        id of object can be empty
      operationId: createLabel
      parameters:
        - $ref: "#/components/parameters/authParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Label"
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          label_id:
                            type: string
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /labels/{label_id}:
    put:
      description: update a Label by its id
      operationId: updateLabel
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: label_id
          in: path
          description: ID of Label to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Label"
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          label_id:
                            type: string
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      description: delete a Label by its id
      operationId: deleteLabel
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: label_id
          in: path
          description: ID of Label to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          label_id:
                            type: string
                            format: uuid
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /annotations:
    get:
      description: get all annotations by queried params
      operationId: fetchAnnotations
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: object_id
          in: query
          description: id of Object that Annotation belongs to
          required: false
          schema:
            type: string
            format: uuid
        - name: study_instance_uid
          in: query
          description: id of Study that Annotation belongs to
          required: false
          schema:
            type: string
            format: uuid
        - name: series_instance_uid
          in: query
          description: id of Series that Annotation belongs to
          required: false
          schema:
            type: string
            format: uuid
        - name: sop_instance_uid
          in: query
          description: id of Instance that Annotation belongs to
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: get annotations response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/AnnotationExtend"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      description: create new Annotation
      deprecated: true
      operationId: createAnnotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Annotation"
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          label_id:
                            type: string
                            format: uuid
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /annotations/{annotation_id}:
    put:
      description: update a Annotation by its id
      operationId: updateAnnotation
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: annotation_id
          in: path
          description: ID of Annotation to fetch
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Annotation"
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          annotation_id:
                            type: string
                            format: uuid
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      description: delete a Annotation by its id
      operationId: deleteAnnotation
      parameters:
        - name: annotation_id
          in: path
          description: delte an Annotation by its id
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/authParam"
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          annotation_id:
                            type: string
                            format: uuid
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /tasks/{task_id}/annotations:
    put:
      description: set many Annotations and Task comment
      operationId: setAnnotations
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                annotations:
                  type: array
                  items:
                    $ref: "#/components/schemas/Annotation"
                comment:
                  type: string

      responses:
        "200":
          description: update task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /label_groups/{label_group_id}/labels:
    post:
      parameters:
        - $ref: "#/components/parameters/authParam"
      description: import labels to label group
      operationId: importLabels
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      parameters:
        - $ref: "#/components/parameters/authParam"
      description: export labels from label group
      operationId: exportLabels
      responses:
        "200":
          description: response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    authParam:
      name: Authorization
      in: header
      schema:
        type: string
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          enum:
            - value: 0
              name: Success
            - value: -1
              name: Failed
            - value: -2
              name: Invalid Data
            - value: -3
              name: Server Error
        message:
          type: string
        count:
          type: integer
          format: int32
          description: size of returned data
    Annotation:
      type: object
      required:
        - label_id
        - label_ids
        - object_id
        - task_id
      properties:
        id:
          type: string
          format: uuid
        label_id:
          deprecated: true
          type: string
          format: uuid
        label_ids:
          type: array
          items:
            type: string
            format: uuid
        object_id:
          type: string
          format: uuid
        creator_id:
          type: string
          format: uuid
        creator_name:
          type: string
        project_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        description:
          type: string
        data:
          anyOf:
            - type: array
              items:
                $ref: "#/components/schemas/Point2D"
            - $ref: "#/components/schemas/Media"
        type:
          type: string
          enum: [BOUNDING_BOX, MASK, POLYGON, TAG]
        meta:
          type: object
          properties:
            study_instance_uid:
              type: string
              format: uuid
            series_instance_uid:
              type: string
              format: uuid
            sop_instance_uid:
              type: string
              format: uuid
    Label:
      type: object
      required:
        - name
        - scope
        - color
        - label_group_id
      properties:
        id:
          type: string
          format: uuid
        parent_label_id:
          type: string
          format: uuid
        name:
          type: string
        tag:
          type: string
        type:
          type: string
          enum: [IMPRESSION, FINDING]
        scope:
          type: string
          enum: [STUDY, SERIES, IMAGE]
        short_name:
          type: string
        creator_id:
          type: string
          format: uuid
        description:
          type: string
        color:
          type: string
        annotation_type:
          type: string
          enum: [BOUNDING_BOX, MASK, POLYGON, TAG]
        label_group_id:
          type: string
          format: uuid
    AnnotationExtend:
      allOf:
        - $ref: "#/components/schemas/Annotation"
        - properties:
            label:
              $ref: "#/components/schemas/Label"
...
    Point2D:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
    Media:
      type: object
      properties:
        mime_type:
          type: string
        bytes:
          type: string
...